% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_one_file}
\alias{parse_one_file}
\title{Parse Summary File}
\usage{
parse_one_file(path)
}
\arguments{
\item{path}{A character string representing the file path to the `.txt` file.}
}
\value{
A data frame with parsed data, analyte labels, and source file name.
}
\description{
Reads and parses a single summary `.txt` file exported from TargetLynx.
}
\examples{
parse_one_file("compound_123.txt")
}
